+++ packages/snap/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `signProofOfOwnership` client request method for silently signing `metamask:proof-of-ownership:<nonce>:<address>` messages with the account's ed25519 key
+- Add `signProofOfOwnership` client request method ([#611](https://github.com/MetaMask/snap-solana-wallet/pull/611))
+  - This method silently signs `metamask:proof-of-ownership:<nonce>:<address>` messages with the account's ed25519 key
 
 ### Changed